+++ fe/fe-core/src/main/java/org/apache/doris/qe/DdlExecutor.java
@@ -25,7 +25,6 @@ import org.apache.doris.analysis.CreateRoutineLoadStmt;
 import org.apache.doris.analysis.CreateTableStmt;
 import org.apache.doris.analysis.DdlStmt;
 import org.apache.doris.analysis.DropUserStmt;
-import org.apache.doris.analysis.RecoverDbStmt;
 import org.apache.doris.analysis.RefreshDbStmt;
 import org.apache.doris.analysis.RefreshTableStmt;
 import org.apache.doris.analysis.SetUserPropertyStmt;
@@ -65,9 +64,6 @@ public class DdlExecutor {
             env.getAuth().dropUser(stmt);
         } else if (ddlStmt instanceof SetUserPropertyStmt) {
             env.getAuth().updateUserProperty((SetUserPropertyStmt) ddlStmt);
-
-        } else if (ddlStmt instanceof RecoverDbStmt) {
-            env.recoverDatabase((RecoverDbStmt) ddlStmt);
         } else if (ddlStmt instanceof SyncStmt) {
             return;
         } else if (ddlStmt instanceof UninstallPluginStmt) {
